Change Armbian v23.11 to v24.11, Add support RepkaPi4 Optimal Support

+++ lib/functions/rootfs/distro-specific.sh
@@ -11,65 +11,39 @@ function install_distribution_specific() {
 	display_alert "Applying distribution specific tweaks for" "${RELEASE:-}" "info"
 
 	# disable broken service, the problem is in default misconfiguration
-	# disable hostapd as it needs to be configured to start correctly
-	disable_systemd_service_sdcard smartmontools.service smartd.service hostapd.service
+	disable_systemd_service_sdcard smartmontools.service smartd.service
 
-	case "${RELEASE}" in
+	if [[ "${DISTRIBUTION}" == "Ubuntu" ]]; then
 
-		focal | jammy | kinetic | lunar | mantic)
+		# by using default lz4 initrd compression leads to corruption, go back to proven method
+		# @TODO: rpardini: this should be a config option (which is always set to zstd ;-D )
+		sed -i "s/^COMPRESS=.*/COMPRESS=gzip/" "${SDCARD}"/etc/initramfs-tools/initramfs.conf
 
-			# by using default lz4 initrd compression leads to corruption, go back to proven method
-			# @TODO: rpardini: this should be a config option (which is always set to zstd ;-D )
-			sed -i "s/^COMPRESS=.*/COMPRESS=gzip/" "${SDCARD}"/etc/initramfs-tools/initramfs.conf
+		run_host_command_logged rm -f "${SDCARD}"/etc/update-motd.d/{10-uname,10-help-text,50-motd-news,80-esm,80-livepatch,90-updates-available,91-release-upgrade,95-hwe-eol}
 
-			run_host_command_logged rm -f "${SDCARD}"/etc/update-motd.d/{10-uname,10-help-text,50-motd-news,80-esm,80-livepatch,90-updates-available,91-release-upgrade,95-hwe-eol}
+		# Journal service adjustements
+		sed -i "s/#Storage=.*/Storage=volatile/g" "${SDCARD}"/etc/systemd/journald.conf
+		sed -i "s/#Compress=.*/Compress=yes/g" "${SDCARD}"/etc/systemd/journald.conf
+		sed -i "s/#RateLimitIntervalSec=.*/RateLimitIntervalSec=30s/g" "${SDCARD}"/etc/systemd/journald.conf
+		sed -i "s/#RateLimitBurst=.*/RateLimitBurst=10000/g" "${SDCARD}"/etc/systemd/journald.conf
 
-			# DNS fix
-			if [[ -n "$NAMESERVER" ]]; then
-				if [[ -f "${SDCARD}"/etc/systemd/resolved.conf ]]; then
-					sed -i "s/#DNS=.*/DNS=$NAMESERVER/g" "${SDCARD}"/etc/systemd/resolved.conf
-				else
-					display_alert "DNS fix" "/etc/systemd/resolved.conf not found: ${DISTRIBUTION} ${RELEASE}" "info"
-				fi
-			fi
+		# disable conflicting services
+		disable_systemd_service_sdcard ondemand.service
 
-			# Journal service adjustements
-			sed -i "s/#Storage=.*/Storage=volatile/g" "${SDCARD}"/etc/systemd/journald.conf
-			sed -i "s/#Compress=.*/Compress=yes/g" "${SDCARD}"/etc/systemd/journald.conf
-			sed -i "s/#RateLimitIntervalSec=.*/RateLimitIntervalSec=30s/g" "${SDCARD}"/etc/systemd/journald.conf
-			sed -i "s/#RateLimitBurst=.*/RateLimitBurst=10000/g" "${SDCARD}"/etc/systemd/journald.conf
-
-			# Chrony temporal fix https://bugs.launchpad.net/ubuntu/+source/chrony/+bug/1878005
-			[[ -f "${SDCARD}"/etc/default/chrony ]] && sed -i '/DAEMON_OPTS=/s/"-F -1"/"-F 0"/' "${SDCARD}"/etc/default/chrony
-
-			# disable conflicting services
-			disable_systemd_service_sdcard ondemand.service
-
-			# Remove Ubuntu APT spamming
-			install_artifact_deb_chroot "fake-ubuntu-advantage-tools"
-			truncate --size=0 "${SDCARD}"/etc/apt/apt.conf.d/20apt-esm-hook.conf
-
-			;;
-	esac
+		# Remove Ubuntu APT spamming
+		install_artifact_deb_chroot "fake-ubuntu-advantage-tools"
+		truncate --size=0 "${SDCARD}"/etc/apt/apt.conf.d/20apt-esm-hook.conf
+	fi
 
 	# install our base-files package (this replaces the original from Debian/Ubuntu)
 	if [[ "${KEEP_ORIGINAL_OS_RELEASE:-"no"}" != "yes" ]]; then
-		install_artifact_deb_chroot "armbian-base-files"
+		install_artifact_deb_chroot "armbian-base-files" "--allow-downgrades"
 	fi
 
-	# Basic Netplan config. Let NetworkManager/networkd manage all devices on this system
-	if [[ -d "${SDCARD}"/etc/netplan ]]; then
-
-		declare RENDERER=networkd
-		if [ -d "${SDCARD}"/etc/NetworkManager ]; then
-			local RENDERER=NetworkManager
-		fi
-
-		cat <<- EOF > "${SDCARD}"/etc/netplan/armbian-default.yaml
-		network:
-		  version: 2
-		  renderer: ${RENDERER}
-		EOF
+	# Set DNS server if systemd-resolved is in use
+	if [[ -n "$NAMESERVER" && -f "${SDCARD}"/etc/systemd/resolved.conf ]]; then
+		display_alert "Using systemd-resolved" "for DNS management" "info"
+		# This used to set a default DNS entry from $NAMESERVER into "${SDCARD}"/etc/systemd/resolved.conf.d/00-armbian-default-dns.conf -- no longer; better left to DHCP.
 	fi
 
 	# cleanup motd services and related files
@@ -95,7 +69,7 @@ function install_distribution_specific() {
 # create_sources_list_and_deploy_repo_key <when> <release> <basedir>
 #
 # <when>: rootfs|image
-# <release>: bullseye|bookworm|sid|focal|jammy|kinetic|lunar|mantic
+# <release>: bullseye|bookworm|sid|focal|jammy|noble|oracular
 # <basedir>: path to root directory
 #
 function create_sources_list_and_deploy_repo_key() {
@@ -112,9 +86,6 @@ function create_sources_list_and_deploy_repo_key() {
 
 				deb http://${DEBIAN_MIRROR} ${release}-updates main contrib non-free
 				#deb-src http://${DEBIAN_MIRROR} ${release}-updates main contrib non-free
-
-				deb http://${DEBIAN_MIRROR} ${release}-backports main contrib non-free
-				#deb-src http://${DEBIAN_MIRROR} ${release}-backports main contrib non-free
 
 				deb http://${DEBIAN_SECURTY} ${release}/updates main contrib non-free
 				#deb-src http://${DEBIAN_SECURTY} ${release}/updates main contrib non-free
@@ -154,13 +125,11 @@ function create_sources_list_and_deploy_repo_key() {
 			EOF
 			;;
 
-		sid) # sid is permanent unstable development and has no such thing as updates or security
+		sid | unstable) # sid is permanent unstable development and has no such thing as updates or security
 			cat <<- EOF > "${basedir}"/etc/apt/sources.list
 				deb http://${DEBIAN_MIRROR} $release main contrib non-free non-free-firmware
 				#deb-src http://${DEBIAN_MIRROR} $release main contrib non-free non-free-firmware
 
-				deb http://${DEBIAN_MIRROR} unstable main contrib non-free non-free-firmware
-				#deb-src http://${DEBIAN_MIRROR} unstable main contrib non-free non-free-firmware
 			EOF
 
 			# Exception: with riscv64 not everything was moved from ports
@@ -170,7 +139,7 @@ function create_sources_list_and_deploy_repo_key() {
 			fi
 			;;
 
-		focal | jammy | kinetic | lunar | mantic)
+		focal | jammy | noble | oracular)
 			cat <<- EOF > "${basedir}"/etc/apt/sources.list
 				deb http://${UBUNTU_MIRROR} $release main restricted universe multiverse
 				#deb-src http://${UBUNTU_MIRROR} $release main restricted universe multiverse
@@ -222,8 +191,10 @@ function create_sources_list_and_deploy_repo_key() {
 	# replace local package server if defined. Suitable for development
 	[[ -n $LOCAL_MIRROR ]] && echo "deb ${SIGNED_BY}http://$LOCAL_MIRROR $RELEASE ${components[*]}" > "${basedir}"/etc/apt/sources.list.d/armbian.list
 
-	# disable repo if SKIP_ARMBIAN_REPO==yes, or if when==image-early.
-	if [[ "${when}" == "image-early" || "${SKIP_ARMBIAN_REPO}" == "yes" ]]; then
+	# disable repo if DISTRIBUTION_STATUS==eos, or if SKIP_ARMBIAN_REPO==yes, or if when==image-early.
+	if [[ "${when}" == "image-early" ||
+		"$(cat "${SRC}/config/distributions/${RELEASE}/support")" == "eos" ||
+		"${SKIP_ARMBIAN_REPO}" == "yes" ]]; then
 		display_alert "Disabling Armbian repo" "${ARCH}-${RELEASE} :: skip:${SKIP_ARMBIAN_REPO:-"no"} when:${when}" "info"
 		mv "${SDCARD}"/etc/apt/sources.list.d/armbian.list "${SDCARD}"/etc/apt/sources.list.d/armbian.list.disabled
 	fi
